[#1250] Compiling 4.2.0 on MacOSX
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@9958 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ common/platform/windows/SysFile.cpp
@@ -854,7 +854,7 @@ bool SysFile::setPosition(int64_t location, int64_t &position)
 
 
     // is this location within the buffer bounds?
-    if (location >= (filePointer - bufferedInput) && location < filePointer)
+    if (location >= (int64_t)(filePointer - bufferedInput) && location < filePointer)
     {
         // just shift the buffer position;
         bufferPosition = (size_t)(location - (filePointer - (int64_t)bufferedInput));
